refactor, clean up code moved most files methods to FilesService
