+++ db-api/src/main/kotlin/com/vsharkovski/dbpaperapi/service/TempDirectoryService.kt
@@ -2,15 +2,13 @@ package com.vsharkovski.dbpaperapi.service
 
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
-import org.springframework.util.FileSystemUtils
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
 import javax.annotation.PostConstruct
-import kotlin.io.path.exists
 
 @Service
-class TempDirectoryService {
+class TempDirectoryService(val filesService: FilesService) {
     @Value("\${temp.path}")
     val tempDirectoryPathString = ""
 
@@ -23,19 +21,6 @@ class TempDirectoryService {
         Files.createDirectories(tempDirectoryPath)
 
         // Clear it.
-        clearDirectory(tempDirectoryPath)
-    }
-
-    fun clearDirectory(directory: Path) {
-        if (!directory.exists()) {
-            return
-        }
-
-        // Delete all subdirectories and files.
-        Files.newDirectoryStream(directory).use { children ->
-            children.forEach {
-                FileSystemUtils.deleteRecursively(it)
-            }
-        }
+        filesService.clearDirectory(tempDirectoryPath)
     }
 }
